pyproject.toml
[tool.poetry]
name = "atomic-habits-spa"
version = "0.1.0"
description = ""
authors = ["vadim marakushin <marakushin.vadim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
ipython = "^8.26.0"
drf-yasg = "^1.21.7"
redis = "^5.0.7"
billiard = "^4.2.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
setuptools = "^70.3.0"
requests = "^2.32.3"
coverage = "^7.6.0"
flake8 = "^7.1.0"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
